fix: move data from component to a data folder
moved the  sidebar data's from the component to its own sidebar.ts file
